refactor: x5c-validator
