--- waltid-libraries/crypto/waltid-crypto/src/commonMain/kotlin/id/walt/crypto/utils/X5CUtils.kt
@@ -1,5 +1,0 @@
-package id.walt.crypto.utils
-
-expect object X5CUtils {
-    fun verifyX5Chain(certificates: List<String>, trustedCA: List<String>): Boolean
-}
