--- waltid-libraries/crypto/waltid-crypto/src/iosMain/kotlin/id/walt/crypto/utils/X5CUtils.ios.kt
@@ -1,5 +1,0 @@
-package id.walt.crypto.utils
-
-actual object X5CUtils {
-    actual fun verifyX5Chain(certificateChain: List<String>, trustedRootCA: List<String>): Boolean = TODO()
-}
