--- waltid-libraries/crypto/waltid-crypto/src/jsMain/kotlin/id/walt/crypto/utils/X5CUtils.js.kt
@@ -1,5 +1,0 @@
-package id.walt.crypto.utils
-
-actual object X5CUtils {
-    actual fun verifyX5Chain(certificates: List<String>, trustedCA: List<String>): Boolean = TODO()
-}
